+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.0-0 / Unreleased
   * [ADD] .set() creates intermediate objects if missing
   * [ADD] .writeToForm() and .readFromForm() accept inputs with nested names
+  * [FIX] .remove() supports dot notation
 
 # 0.0.1 / 2016-04-10
   * Initial release