+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# 1.0.0-0 / Unreleased
+# 1.1.0-0 / Unreleased
+  * [CHANGE] rename `.init` options to `.defaults`
+
+# 1.0.0 / 2016-08-11
   * [ADD] .set() creates intermediate objects if missing
   * [ADD] .writeToForm() and .readFromForm() accept inputs with nested names
   * [FIX] .remove() supports dot notation